+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/postgresql/PostgreSQLStorageConfig.java
@@ -25,16 +25,4 @@ import org.apache.skywalking.oap.server.storage.plugin.jdbc.mysql.MySQLStorageCo
 @Setter
 @Getter
 public class PostgreSQLStorageConfig extends MySQLStorageConfig {
-    /**
-     * The maximum size of batch size of SQL execution
-     *
-     * @since 8.8.0
-     */
-    private int maxSizeOfBatchSql = 2000;
-    /**
-     * async batch execute pool size
-     *
-     * @since 8.8.0
-     */
-    private int asyncBatchPersistentPoolSize  = 4;
 }